add more descriptive getting started
Rewrote the getting started, because I had trouble grasping what I should do when starting out, and I already know about Pester and Powershell. The upper part in readme now describes what pskoans is in general, and links to the bottom where a full guide is. Personally I would also move the repo installation under the getting started, because almost everyone uses psgallery now, but did not want to change that much. I also changed the first koan, because the readme on the top of the file mentions that __ should be replaced, yet the first example did not use it..