Fixes auto-succeeding pester tests in AboutComparison and AboutStrings (#214)
* Fixed "adds strings together" 

It "adds strings together" was passing the pester test without user input. Replaced the contents of $String1 with '__' so the user can solve it.

* Fixed It's that are autopassing the pester tests

Auto-passing tests:
It 'is a simple test'
It 'returns $true if either input is $true'
It 'negates a boolean value'
In all three instances, __ evaluates to $true, which causes the tests to pass. 

I replaced __ with $__ in these instances to make sure the tests fail.

* Updated $__ to $____ 

Added two extra underscores to each $__ to make sure no one will confuse $__ for $_

* Update PSKoans/Koans/Foundations/AboutStrings.Koans.ps1

Longer blanks! Better blanks!

Co-Authored-By: Joel Sallow (/u/ta11ow) <32407840+vexx32@users.noreply.github.com>