:sparkles: Add topic data caching for -Meditate
Adds functionality to store topic completion data to a module-scope var
with Get-Karma.
This simple cache will then be pulled and subsequently cleared by
Show-Karma when used with -Contemplate.

If there is no cached data, Show-Karma will call Get-Karma.

The cache allows us to avoid double-calling Get-Karma when, for example,
the user calls Show-Karma to view results and then follows with
Show-Karma -Contemplate to open the file.