(#413) Refactor Measure-Koan to use SkipRun
Pester v5.2.0 added the ability to natively run just the discovery phase
of a test suite, so we refactor Measure-Koan to utilize this and save
doing a lot of manual work.

This also obsoletes the private Measure-KoanTestBlock command that
previously supported Measure-Koan. Pester automatically does that
calculation on its own before presenting the discovery results. This
command has been removed as a result.