:rocket: (maint) Ensure build publishes Pester
In some cases it looks like Pester can fail to publish if there are too
many versions of Pester available. Prevent this by checking the version
we actually need and publishing just that version when doing a publish
to the filesystem.

Also set the initially installed version of pester to the correct
version based on the current required version specified by PSKoans
itself, instead of having to update the environment-setup.yml each time.